Move tests/ directory out of ruddock/
- Have the tests/ directory be a top level directory, with a directory
  structure that matches ruddock/
- Remove the db fixture and fix tests.
- Fix the gitignore with an update python gitignore (that ignores
  .cache now).
