+++ pytest.ini
@@ -1,1 +1,3 @@
 [pytest]
+# All tests should be in the test/ directory.
+testpaths = tests
