+++ tests/admin/test_admin_routes.py
@@ -7,8 +7,8 @@ import httplib
 
 from ruddock import auth_utils
 from ruddock.resources import Permissions
-from ruddock.tests import utils
-from ruddock.tests.fixtures import client
+from tests import utils
+from tests.fixtures import client
 
 def test_admin_home(client):
   """Tests the /admin route."""
