+++ tests/users/test_users_routes.py
@@ -1,0 +1,26 @@
+"""
+Tests routes in the users module.
+"""
+
+import flask
+import sqlalchemy
+import httplib
+
+from tests import utils
+from tests.fixtures import client
+
+def test_show_memberlist(client):
+  """Tests /members route."""
+  with client.session_transaction() as session:
+    utils.login(session)
+  response = client.get(flask.url_for('users.show_memberlist'))
+  assert response.status_code == httplib.OK
+
+def test_view_profile(client):
+  """Tests that viewing a user works."""
+  username = "twilight"
+  with client.session_transaction() as session:
+    utils.login(session)
+  response = client.get(
+      flask.url_for('users.view_profile', username=username))
+  assert response.status_code == httplib.OK
